feat: moved database to a standalone class
M  src/main/java/com/tuvarna/phd/controller/BaseController.java
R  src/main/java/com/tuvarna/phd/models/ControllerResponse.java -> src/main/java/com/tuvarna/phd/controller/ControllerResponse.java
A  src/main/java/com/tuvarna/phd/model/DatabaseModel.java
R  src/main/java/com/tuvarna/phd/models/S3Model.java -> src/main/java/com/tuvarna/phd/model/S3Model.java
M  src/main/java/com/tuvarna/phd/service/AuthServiceImpl.java
